v3.8.0-rc.6: round-23 audit response — 6 fixes (T-FLAKE-1, N-4, N-5, ARCH-1, R-4, OIA→required)
Fix 1 T-FLAKE-1: add vitest per-it timeout 90s to build-embeddings test; prevents
the 15s global testTimeout from racing the 60s spawnSync subprocess cold-start.

Fix 2 N-4: npm audit fix — protobufjs 7.5.7 → 7.6.0 (GHSA-jggg-4jg4-v7c6,
moderate DoS via unbounded recursive JSON descriptor expansion).

Fix 3 N-5: serve-http --watch help text updated to mention .pdf and embed-db
re-sync (was ".md changes only" — inconsistent with api.md since rc.2/rc.3).

Fix 4 ARCH-1: break circular import embed-pipeline → server → embed-pipeline.
Move buildEmbedText from server.ts to embed-pipeline.ts; server.ts re-exports
it for backward compat. TSDoc preserved + migration note added in same commit.

Fix 5 R-4: contextPack hard budget cap. Per-section checks leave systematic
gaps (section headers, join overhead). Hard-slice raw bundle to charBudget
with "[…budget cap reached…]" marker as defense-in-depth.

Fix 6 CI: promote check:oia from advisory → required (9th gate). Zero false
positives across 6 releases since v3.7.19. REQUIRED regex + REQ_COUNT + README
+ CLAUDE.md all updated. docs-consistency invariant caught the REQ_COUNT drift
immediately during local test run (exactly as designed).

838 tests (unchanged). All 9 required CI gates pass locally.
