+++ CLAUDE.md
@@ -28,7 +28,7 @@ Directive: **"Максимальное качество и уверенный т
 4. Coverage thresholds met (lines ≥86, statements ≥82, functions ≥75, branches ≥74)
 5. `scripts/check-changelog-coverage.mjs` passes (CHANGELOG claims = reality within 0.5pp)
 6. `scripts/smoke.mjs` synthetic vault scan passes
-7. All 8 required CI gates green on PR (lint, test×2, smoke, audit, coverage, version-consistency, docs — `docs` added v3.7.10, count locked at 8 since v3.7.13)
+7. All 9 required CI gates green on PR (lint, test×2, smoke, audit, coverage, version-consistency, docs, oia — `docs` added v3.7.10, `oia` promoted from advisory v3.8.0-rc.6)
 8. Daily-check report after merge: 0 regressions, 0 new CodeQL / Dependabot alerts
 9. CHANGELOG entry with TL;DR blockquote + method note
 10. All docs-consistency invariants stay green (extend as new surfaces appear)
@@ -108,8 +108,9 @@ Apply **root-cause sweep methodology** consistently: every bug → identify the 
 - **v3.7.13 shipped**: round-15 external audit response — 4th independent external audit since v3.6.0. 15 ship-ready findings: H1 PDF page slicing pre-extraction DoS, H2 HTTP stateful pre-initialize resource leak, H3 engines.node floor mismatch, M1 renameNote rollback recovery, M2 overwrite=false atomic create (wx flag), M3 OCR lang regex validation, M5 7→8 required gates, M8 per-file coverage fail-on-missing, M9 chmod parentExisted pattern in saveDiskCache, M10 docs/api.md broken link → SECURITY.md, M11 benchmarks latency dedup, M12 COMPARISON test count gate, L2 OCR/read_pdf page range refine, L4 github-metadata fail-loud + CI GH_TOKEN, L7 exclude docs/audits from npm package. 2 findings already closed by v3.7.12 (L1 `.base` normalization, L3 canvas mtimeMs). **813 tests** (+11 negative-controls); +1 docs-consistency invariant; npm tarball -228 KB (docs/audits/ excluded).
 - **v3.7.14 shipped**: round-16 SELF-audit — class-sweep methodology applied to v3.7.13 fixes. 5 findings: F1 v3.7.13 silent overclaim (M1 fixed renameNote ordering but TSDoc header still described OLD buggy order — 6th overclaim instance), F2 M2 sibling (renameFile had same stat-then-rename TOCTOU race; fix via link()+unlink() atomic exclusive create), F3 appendNote stat-then-append race let parallel writers bypass maxFileBytes (fix via single open fd + write), F4 hardcoded "8 required CI gates" not gated by docs-consistency (v3.5.9 anti-pattern recurrence — new invariant against release.yml REQUIRED regex), F5 automated gh release create in release.yml (M13 from round-15, originally v3.8.0-deferred). **815 tests** (+2 negative-controls + invariant); +1 docs-consistency invariant.
 - **v3.7.15 shipped**: round-17 POST-MERGE audit on v3.7.14 — **meta-recursion finding**: v3.7.14 F1 closed overclaim #6 but v3.7.14 F2 SHIPPED overclaim #7 in the same patch (renameFile TSDoc still said "Atomic via fs.rename" after F2 changed the impl to link()+unlink()). 3 fixes (R17-1 renameFile TSDoc, R17-2 appendNote TSDoc enhancement, R17-3 docs/COMPARISON.md "5 models" missed v3.7.12 L4 instance) + 1 new invariant (R17-4 COMPARISON reranker honesty). Plus overclaim #8 (v3.7.14 orphan-tag procedural error caught by "Assert tag is on main" guard) + overclaim #9 (v3.7.14 F5 was permission-incomplete, `HTTP 403`, fixed via `permissions: contents: write`). 9 documented overclaim instances now. CLAUDE.md anti-patterns extended with 3 new rules. **816 tests** (+1 invariant). F5 release automation finally verified working end-to-end on v3.7.15 tag push.
-- **v3.7.16 shipped (current)**: round-18 external audit response — 5th independent external audit since v3.6.0, on the v3.7.5 codebase (commit b9daf39). The v3.7.6→v3.7.15 cascade had already closed many findings; this patch addresses still-open critical + high-impact items. **10 fixes**: P1-1 OCR network-policy disclosure (stderr warning + SECURITY.md), P1-2 OCR 200-page default cap, P1-4 persistent-cache + privacy filter lifecycle, P1-5 watcher PDF lifecycle, P1-6 macOS case-insensitive write privacy bypass (canonicalize parent via realpath), P2-13 title-based write fail-on-ambiguity (silent data corruption), P2-14 validateNoteProposal privacy check, P2-15 FTS5 tag LIKE `%`/`_` escape, P2-16 graph_boost `#chunk-N` regex (fixes `C# Notes.md`), P3-23 PDF install hint v5.7.284, P3-28 safeFts5Query reserved-word quoting (contract change), P3-30 issue template ChatGPT + Claude Desktop. 14 architectural findings deferred to v3.8.0 backlog. **816 tests** (1 test contract change for P3-28).
-- **v3.7+ backlog status**: All known audit findings from rounds 1–18 closed. v3.8.0 backlog: serve-http parity (P1-3), OCR concurrency/timeout (#2 beyond 200-page cap), search underfill (P2-7), FTS5/embedding chunking parity (P2-8), stateful session race (P2-10), HTTP server close cleanup (P2-11), doctor privacy filters (P2-12), npm package broken links (P2-18), canonical CLI docs completeness (P2-20), FTS gating wording (P3-21), heading tilde fences (P3-25), substring search contract (P3-26), HNSW metadata validation (P3-27), setup-snippet mkdir (P3-29) + round-14/15 originals. **No open audit items as of 2026-05-18 / 26 releases into the v3.6.0 → v3.7.16 cascade.**
+- **v3.7.16 shipped**: round-18 external audit response — 5th independent external audit since v3.6.0, on the v3.7.5 codebase (commit b9daf39). The v3.7.6→v3.7.15 cascade had already closed many findings; this patch addresses still-open critical + high-impact items. **10 fixes**: P1-1 OCR network-policy disclosure (stderr warning + SECURITY.md), P1-2 OCR 200-page default cap, P1-4 persistent-cache + privacy filter lifecycle, P1-5 watcher PDF lifecycle, P1-6 macOS case-insensitive write privacy bypass (canonicalize parent via realpath), P2-13 title-based write fail-on-ambiguity (silent data corruption), P2-14 validateNoteProposal privacy check, P2-15 FTS5 tag LIKE `%`/`_` escape, P2-16 graph_boost `#chunk-N` regex (fixes `C# Notes.md`), P3-23 PDF install hint v5.7.284, P3-28 safeFts5Query reserved-word quoting (contract change), P3-30 issue template ChatGPT + Claude Desktop. 14 architectural findings deferred to v3.8.0 backlog. **816 tests** (1 test contract change for P3-28).
+- **v3.8.0-rc.6 shipped (current)**: round-23 external audit response (6 fixes). T-FLAKE-1 vitest per-it timeout (build-embeddings CI flake), N-4 protobufjs GHSA-jggg-4jg4-v7c6, N-5 serve-http `--watch` help parity, ARCH-1 circular import (`buildEmbedText` → embed-pipeline.ts), R-4 contextPack hard budget cap, OIA promoted advisory → required (9th gate). **838 tests** (unchanged). No open audit items from rounds 1–23 as of 2026-05-21.
+- **v3.8.0 backlog status**: All known audit findings from rounds 1–23 closed. Remaining v3.8.0 items: serve-http parity (P1-3), OCR concurrency/timeout, search underfill (P2-7), FTS5/embedding chunking parity (P2-8), stateful session race (P2-10), HTTP server close cleanup (P2-11), doctor privacy filters (P2-12), npm package broken links (P2-18), canonical CLI docs (P2-20), FTS gating wording (P3-21), heading tilde fences (P3-25), HNSW metadata validation (P3-27), setup-snippet mkdir (P3-29) + other v3.8.0 backlog originals. External audit before promotion to v3.8.0 stable.
 - **v3.7+ deferred** (no audit pressure, parked for v3.8+):
   - E2E preservation tests for `setup` / `eval` / `build-embeddings` (currently only `index` has E2E preservation+forced-rebuild pair from v3.6.4).
   - Strengthen K-1 invariant via TypeScript AST: enforce that peek result is CONSUMED in the constructor's `modelAlias` / `tokenize` arg, not just present in scope.
